handle quant mixtral with care
