+++ aphrodite/modeling/loader.py
@@ -4,7 +4,6 @@ from typing import Optional, Type
 
 import torch
 import torch.nn as nn
-from transformers import PretrainedConfig
 
 from aphrodite.common.config import DeviceConfig, ModelConfig, LoRAConfig
 from aphrodite.modeling.models import ModelRegistry
@@ -23,9 +22,11 @@ def _set_default_torch_dtype(dtype: torch.dtype):
 
 def _get_model_architecture(model_config: ModelConfig) -> Type[nn.Module]:
     architectures = getattr(model_config.hf_config, "architectures", [])
-    # FIXME: This is a hack.
-    if (model_config.quantization is None and "MixtralForCausalLM" in architectures):
-        architectures = ["FusedMixtralForCausalLM"]
+    # Special handling for quantized Mixtral.
+    # FIXME: This is a temporary hack.
+    if (model_config.quantization is not None
+            and "MixtralForCausalLM" in architectures):
+        architectures = ["QuantMixtralForCausalLM"]
     for arch in architectures:
         model_cls = ModelRegistry.load_model_cls(arch)
         if model_cls is not None:
